Renamed userscript_api::queue to queue_api
