+++ src/userscript_api.rs
@@ -24,7 +24,7 @@
 //!
 
 pub mod help_system;
-pub mod queue;
+pub mod queue_api;
 pub mod user_engine;
 pub mod include {
     //! # Useful re-exports from other crates.
